+++ lib/src/actions/cube_action.dart
@@ -1,0 +1,16 @@
+export 'navigator_action.dart';
+
+///
+/// Created by
+///
+/// ─▄▀─▄▀
+/// ──▀──▀
+/// █▀▀▀▀▀█▄
+/// █░░░░░█─█
+/// ▀▄▄▄▄▄▀▀
+///
+/// Rafaelbarbosatec
+/// on 14/02/22
+/// Base to Actions
+
+abstract class CubeAction {}